Modified perform action service.

+++ backend/app/db/models/xLogInbound.js
@@ -11,7 +11,7 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER(9).UNSIGNED,
       allowNull: false
     },
-    list_id: DataTypes.INTEGER(14).UNSIGNED,
+    list_id: DataTypes.BIGINT(14).UNSIGNED,
     campaign_id: DataTypes.STRING(20),
     call_date: DataTypes.DATE,
     start_epoch: DataTypes.INTEGER(10).UNSIGNED,
